add line cursor that highlights process info
Now user can easily see all information about the process, and
in the next updates, there'll be a possibility to send signals.
There are still magic numbers in the code responsible for cursor
movement, but they depend solely on starting coordinates of drawing,
and I highly doubt it will change in the future.